Setting apache to run as vagrant
This fixes issues to do with httpd running on unix hosts. For some reason apache can't write to the assets folders and so on and one can't change the access permissions to those folders in the guest.

This gets around the problem by running apache as vagrant (the user that owns the mounted folder)